+++ Changelog.md
@@ -4,6 +4,7 @@
 
 - Support generating an element ID from a Refracted classes.
 - Element IDs will *always* be Refracted elements.
+- Any meta or attributes attached to existing element IDs will no longer be removed.
 
 # 1.1.3 - 2017-03-17